+++ src/main/java/com/team2/fitinside/member/service/AuthService.java
@@ -7,10 +7,10 @@ import com.team2.fitinside.member.dto.MemberResponseDto;
 import com.team2.fitinside.member.dto.TokenDto;
 import com.team2.fitinside.member.entity.Authority;
 import com.team2.fitinside.member.entity.Member;
-import com.team2.fitinside.member.jwt.TokenProvider;
+import com.team2.fitinside.jwt.TokenProvider;
 import com.team2.fitinside.member.mapper.MemberMapper;
-import com.team2.fitinside.member.oath.repository.RefreshTokenRepository;
-import com.team2.fitinside.member.oath.util.RefreshTokenCookieUtil;
+import com.team2.fitinside.oath.repository.RefreshTokenRepository;
+import com.team2.fitinside.oath.util.RefreshTokenCookieUtil;
 import com.team2.fitinside.member.repository.MemberRepository;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
