+++ src/main/java/com/team2/fitinside/oath/dto/CreateAccessTokenResponse.java
@@ -1,4 +1,4 @@
-package com.team2.fitinside.member.oath.dto;
+package com.team2.fitinside.oath.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
