+++ src/main/java/com/team2/fitinside/oath/repository/RefreshTokenRepository.java
@@ -1,6 +1,6 @@
-package com.team2.fitinside.member.oath.repository;
+package com.team2.fitinside.oath.repository;
 
-import com.team2.fitinside.member.oath.entity.RefreshToken;
+import com.team2.fitinside.oath.entity.RefreshToken;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
